refactor: map system and add doxygen to methods/classes that don't have one
