+++ scripts/managers/game_manager/GameManager.cs
@@ -1,7 +1,7 @@
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using AshesOfVelsingrad.Systems;
+using AshesOfVelsingrad.Utilities;
 using Godot;
 
 namespace AshesOfVelsingrad.Managers;
@@ -13,10 +13,10 @@ public partial class GameManager : BaseManager
 {
     #region Private Fields
 
-    private GameOutcome _gameOutcome = GameOutcome.Ongoing;
+    private AovDataStructures.GameOutcome _gameOutcome = AovDataStructures.GameOutcome.Ongoing;
     private bool _isPlayerTurn;
     private bool _unitMoved;
-    private ClickOnMapContext _clickOnMapContext = ClickOnMapContext.MoveUnit;
+    private AovDataStructures.ClickOnMapContext _clickOnMapContext = AovDataStructures.ClickOnMapContext.MoveUnit;
     private readonly List<UnitSystem> _playerUnits = [];
     private readonly List<UnitSystem> _enemyUnits = [];
     private List<(int, int, int)> _currentUnitPossibleMoves = [];
@@ -55,8 +55,8 @@ public partial class GameManager : BaseManager
     #region Private Properties
 
     /// <summary>
-    /// Singleton instance of the <see cref="GameManager"/>.
-    /// Ensures that only one active instance exists at any given time.
+    ///     Singleton instance of the <see cref="GameManager" />.
+    ///     Ensures that only one active instance exists at any given time.
     /// </summary>
     private new static GameManager? Instance { get; set; }
 
@@ -64,7 +64,7 @@ public partial class GameManager : BaseManager
 
     #region Class initialization
 
-    /// <inheritdoc/>
+    /// <inheritdoc />
     public override void _Ready()
     {
         base._Ready();
@@ -95,11 +95,11 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Initializes references to all major systems and sets up initial bindings.
+    ///     Initializes references to all major systems and sets up initial bindings.
     /// </summary>
     /// <remarks>
-    /// This method is called during <see cref="Initialize"/> to connect signals,
-    /// set up turn events, load units, and prepare the map and input systems.
+    ///     This method is called during <see cref="Initialize" /> to connect signals,
+    ///     set up turn events, load units, and prepare the map and input systems.
     /// </remarks>
     private void InitializeGameManager()
     {
@@ -131,7 +131,7 @@ public partial class GameManager : BaseManager
     #region Private Methods
 
     /// <summary>
-    /// Waits for the scene tree to be fully ready, then starts the battle loop.
+    ///     Waits for the scene tree to be fully ready, then starts the battle loop.
     /// </summary>
     /// <returns>A task that completes once the battle has started.</returns>
     private async Task StartBattleWhenReady()
@@ -150,7 +150,7 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Activate player unit (actions and inputs) at the start of their turn.
+    ///     Activate player unit (actions and inputs) at the start of their turn.
     /// </summary>
     private void ActivatePlayerUnit()
     {
@@ -181,7 +181,7 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Disables player unit (actions and inputs) when their turn ends.
+    ///     Disables player unit (actions and inputs) when their turn ends.
     /// </summary>
     private void DeactivatePlayerUnit()
     {
@@ -203,7 +203,7 @@ public partial class GameManager : BaseManager
             return;
         }
 
-        _clickOnMapContext = ClickOnMapContext.MoveUnit;
+        _clickOnMapContext = AovDataStructures.ClickOnMapContext.MoveUnit;
         _isPlayerTurn = false;
         _selectedSkill = null;
         _unitMoved = false;
@@ -215,10 +215,10 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Handles player unit select skill input.
+    ///     Handles player unit select skill input.
     /// </summary>
     /// <remarks>
-    /// Called when <see cref="BattleInputSystem.OnSelectedSkillPressed"/> is triggered.
+    ///     Called when <see cref="BattleInputSystem.OnSelectedSkillPressed" /> is triggered.
     /// </remarks>
     private void PlayerSelectedSkill(int skillId)
     {
@@ -248,7 +248,7 @@ public partial class GameManager : BaseManager
         }
 
         GD.Print($"Selected Skill {skillId}");
-        _clickOnMapContext = ClickOnMapContext.SelectUnitTarget;
+        _clickOnMapContext = AovDataStructures.ClickOnMapContext.SelectUnitTarget;
         _selectedSkill = _turnManagerContainer.GetCurrentUnit().ActiveSkills[skillId];
         _currentUnitReachableCellsForCurrentSelectedSkill = _turnManagerContainer
             .GetCurrentUnit()
@@ -259,10 +259,10 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Handles player unit pass turn input.
+    ///     Handles player unit pass turn input.
     /// </summary>
     /// <remarks>
-    /// Called when <see cref="BattleInputSystem.OnPassTurnPressed"/> is triggered.
+    ///     Called when <see cref="BattleInputSystem.OnPassTurnPressed" /> is triggered.
     /// </remarks>
     private void PlayerPassedUnitTurn()
     {
@@ -278,19 +278,19 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Handles input when the player selects the “Move” action.
+    ///     Handles input when the player selects the “Move” action.
     /// </summary>
     private void PlayerSelectedMove()
     {
         if (_unitMoved)
             return;
         GD.Print("Selected move action.");
-        _clickOnMapContext = ClickOnMapContext.MoveUnit;
+        _clickOnMapContext = AovDataStructures.ClickOnMapContext.MoveUnit;
     }
 
     /// <summary>
-    /// Determines what to do when the player clicks on the map —
-    /// either move a unit or select a skill target depending on the current context.
+    ///     Determines what to do when the player clicks on the map —
+    ///     either move a unit or select a skill target depending on the current context.
     /// </summary>
     /// <param name="cell">The grid cell clicked by the player.</param>
     private void PlayerMovedUnitOrSelectedTarget(Vector3I cell)
@@ -317,17 +317,17 @@ public partial class GameManager : BaseManager
 
         switch (_clickOnMapContext)
         {
-            case ClickOnMapContext.MoveUnit:
+            case AovDataStructures.ClickOnMapContext.MoveUnit:
                 HandlePlayerUnitMove(cell);
                 break;
-            case ClickOnMapContext.SelectUnitTarget:
+            case AovDataStructures.ClickOnMapContext.SelectUnitTarget:
                 HandlePlayerSelectTarget(cell);
                 break;
         }
     }
 
     /// <summary>
-    /// Resets movement flags when the enemy turn ends.
+    ///     Resets movement flags when the enemy turn ends.
     /// </summary>
     private void EnemyTurnEnded()
     {
@@ -343,7 +343,7 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Handles all logic that should occur at the end of the current turn.
+    ///     Handles all logic that should occur at the end of the current turn.
     /// </summary>
     private void CurrentTurnEnded()
     {
@@ -367,19 +367,19 @@ public partial class GameManager : BaseManager
     #region Public Methods
 
     /// <summary>
-    /// Moves the currently active unit to the specified cell on the map.
+    ///     Moves the currently active unit to the specified cell on the map.
     /// </summary>
     /// <param name="cell">
-    /// The target cell position as a <see cref="Vector3I"/> (grid coordinates X, Y, Z).
+    ///     The target cell position as a <see cref="Vector3I" /> (grid coordinates X, Y, Z).
     /// </param>
     /// <remarks>
-    /// Checks that both <c>_mapSystemContainer</c> and <c>_turnManagerContainer</c>
-    /// are properly set before performing the move.
-    /// If valid, the active unit is repositioned in the world based on the given
-    /// cell coordinates and the map’s cell size.
+    ///     Checks that both <c>_mapSystemContainer</c> and <c>_turnManagerContainer</c>
+    ///     are properly set before performing the move.
+    ///     If valid, the active unit is repositioned in the world based on the given
+    ///     cell coordinates and the map’s cell size.
     /// </remarks>
     /// <example>
-    /// <code>
+    ///     <code>
     /// Vector3I targetCell = new Vector3I(2, 0, 3);
     /// gameManager.MoveUnit(targetCell);
     /// </code>
@@ -422,25 +422,25 @@ public partial class GameManager : BaseManager
     }
 
     /// <summary>
-    /// Uses a skill from a source unit on a target unit, depending on the skill’s target type.
+    ///     Uses a skill from a source unit on a target unit, depending on the skill’s target type.
     /// </summary>
     /// <param name="sourceUnit">
-    /// The unit that performs the skill.
+    ///     The unit that performs the skill.
     /// </param>
     /// <param name="targetUnit">
-    /// The main target unit of the skill.
+    ///     The main target unit of the skill.
     /// </param>
     /// <param name="skill">
-    /// The skill being used, containing information about its target type and effects.
+    ///     The skill being used, containing information about its target type and effects.
     /// </param>
     /// <remarks>
-    /// Determines which units are allies or enemies based on the source unit,
-    /// then applies the skill accordingly depending on its <c>TargetType</c>.
-    /// If required containers are not initialized, the method prints an error
-    /// and exits early.
+    ///     Determines which units are allies or enemies based on the source unit,
+    ///     then applies the skill accordingly depending on its <c>TargetType</c>.
+    ///     If required containers are not initialized, the method prints an error
+    ///     and exits early.
     /// </remarks>
     /// <example>
-    /// <code>
+    ///     <code>
     /// SkillSystem healSkill = new SkillSystem("Heal", TargetTypes.SingleAlly);
     /// gameManager.UseSkill(playerUnit, allyUnit, healSkill);
     /// </code>
@@ -467,7 +467,7 @@ public partial class GameManager : BaseManager
 
         switch (skill.TargetType)
         {
-            case TargetTypes.SingleAlly:
+            case AovDataStructures.TargetTypes.SingleAlly:
                 if (!allyUnits.Contains(targetUnit))
                 {
                     GD.PrintErr($"Ally unit {targetUnit.UnitName} not found.");
@@ -481,7 +481,7 @@ public partial class GameManager : BaseManager
                 _turnManagerContainer.GetCurrentUnit().Play(targetUnits, _mapSystemContainer, skill);
                 break;
 
-            case TargetTypes.SingleEnemy:
+            case AovDataStructures.TargetTypes.SingleEnemy:
                 if (!enemyUnits.Contains(targetUnit))
                 {
                     GD.PrintErr($"Enemy unit {targetUnit.UnitName} not found.");
@@ -495,7 +495,7 @@ public partial class GameManager : BaseManager
                 _turnManagerContainer.GetCurrentUnit().Play(targetUnits, _mapSystemContainer, skill);
                 break;
 
-            case TargetTypes.AllAllies:
+            case AovDataStructures.TargetTypes.AllAllies:
                 if (!allyUnits.Contains(targetUnit))
                 {
                     GD.PrintErr($"Ally unit {targetUnit.UnitName} not found.");
@@ -508,7 +508,7 @@ public partial class GameManager : BaseManager
                 _turnManagerContainer.GetCurrentUnit().Play(allyUnits, _mapSystemContainer, skill);
                 break;
 
-            case TargetTypes.AllEnemies:
+            case AovDataStructures.TargetTypes.AllEnemies:
                 if (!enemyUnits.Contains(targetUnit))
                 {
                     GD.PrintErr($"Enemy unit {targetUnit.UnitName} not found.");
