+++ scripts/systems/status_effects/StatusEffectSystem.cs
@@ -1,18 +1,18 @@
 using System.Collections.Generic;
 using System.Linq;
-using AshesOfVelsingrad.utilities;
+using AshesOfVelsingrad.Utilities;
 
 namespace AshesOfVelsingrad.Systems;
 
 /// <summary>
-///     Manages the lifecycle and application of <see cref="StatusEffect{TTarget}"/>s
+///     Manages the lifecycle and application of <see cref="StatusEffect{TTarget}" />s
 ///     for all tracked targets in the game, including units and map cells.
 /// </summary>
 /// <remarks>
 ///     This system keeps track of every target that has active status effects,
 ///     applies new effects, handles stacking, and processes end-of-turn updates.
-///     Targets must implement <see cref="IEffectTarget{TTarget}"/> and be either
-///     a <see cref="UnitSystem"/> or a <see cref="MapSystem"/>.
+///     Targets must implement <see cref="IEffectTarget{TTarget}" /> and be either
+///     a <see cref="UnitSystem" /> or a <see cref="MapSystem" />.
 /// </remarks>
 public sealed class StatusEffectSystem
 {
@@ -27,11 +27,11 @@ public sealed class StatusEffectSystem
     ///     Updates all active status effects on the given target at the end of a turn.
     /// </summary>
     /// <typeparam name="TTarget">
-    ///     The concrete type of the effect target (e.g. <see cref="UnitSystem"/> or
-    ///     <see cref="CellInformation"/>).
+    ///     The concrete type of the effect target (e.g. <see cref="UnitSystem" /> or
+    ///     <see cref="CellInformation" />).
     /// </typeparam>
     /// <param name="target">
-    ///     The target whose active <see cref="StatusEffect{TTarget}"/> instances
+    ///     The target whose active <see cref="StatusEffect{TTarget}" /> instances
     ///     should be updated.
     /// </param>
     private void RefreshTargetEffectsOnTurnEnd<TTarget>(IEffectTarget<TTarget> target)
@@ -59,13 +59,13 @@ public sealed class StatusEffectSystem
     ///     Applies a new status effect to the given target, or stacks it if already present.
     /// </summary>
     /// <typeparam name="TTarget">
-    ///     The type of the target, must be either <see cref="UnitSystem"/> or <see cref="MapSystem"/>.
+    ///     The type of the target, must be either <see cref="UnitSystem" /> or <see cref="MapSystem" />.
     /// </typeparam>
     /// <param name="target">
     ///     The target on which to apply the status effect.
     /// </param>
     /// <param name="newEffect">
-    ///     The <see cref="StatusEffect{TTarget}"/> to apply.
+    ///     The <see cref="StatusEffect{TTarget}" /> to apply.
     /// </param>
     /// <remarks>
     ///     If the target already has an effect of the same type and it is stackable,
@@ -95,7 +95,7 @@ public sealed class StatusEffectSystem
     ///     The target whose status effects should be updated.
     /// </param>
     /// <remarks>
-    ///     Each status effect has its <see cref="StatusEffect{UnitSystem}.OnTurnPassed"/> method called,
+    ///     Each status effect has its <see cref="StatusEffect{UnitSystem}.OnTurnPassed" /> method called,
     ///     and expired effects are removed automatically. If a target has no more active effects,
     ///     it is removed from the tracking list.
     /// </remarks>
@@ -111,10 +111,10 @@ public sealed class StatusEffectSystem
     ///     Processes end-of-turn updates for all tracked targets.
     /// </summary>
     /// <remarks>
-    ///     Iterates over all tracked targets in <see cref="_allTargets"/> and updates
+    ///     Iterates over all tracked targets in <see cref="_allTargets" /> and updates
     ///     their status effects. Expired effects are removed, and targets with no remaining
     ///     effects are removed from the tracking list.
-    ///     Only targets of type <see cref="CellInformation"/> are processed by this method.
+    ///     Only targets of type <see cref="CellInformation" /> are processed by this method.
     /// </remarks>
     public void ProcessTurnEnd()
     {
